+++ scripts/mcp-tool-pack-baseline.json
@@ -12,9 +12,7 @@
   "diagnose_sandbox",
   "edit_sandbox_file",
   "execute_promotion",
-  "generate_code",
   "get_build_dispatch_history",
-  "get_build_engine_readiness",
   "get_build_progress_visibility",
   "get_build_sandbox_state",
   "get_build_scoped_verification",
@@ -24,29 +22,20 @@
   "list_build_activity_since",
   "list_sandbox_files",
   "principle_decide",
-  "promote_to_build_studio",
   "propose_build_decomposition",
   "propose_decomposition",
-  "propose_file_change",
   "propose_improvement",
   "propose_skill_improvement",
-  "provision_build_engine",
   "read_sandbox_file",
-  "reconcile_build_engines",
   "record_decomposition_override",
-  "record_execution_evidence",
   "record_external_development_evidence",
-  "record_functional_failure_evidence",
-  "record_local_integration_result",
   "recover_sandbox",
   "register_digital_product_from_build",
   "reviewBuildPlan",
   "reviewDesignDoc",
-  "run_endpoint_tests",
   "run_release_gate",
   "run_sandbox_command",
   "run_sandbox_tests",
-  "run_tool_script",
   "run_ux_test",
   "saveBuildEvidence",
   "save_build_notes",
@@ -61,9 +50,7 @@
   "start_scout_research",
   "submit_feedback",
   "update_feature_brief",
-  "update_lifecycle",
   "validate_schema",
   "verification_preflight",
-  "verify_live_install_readiness",
   "write_sandbox_file"
 ]
